feat(ToastNotification)!: rename from Toast to ToastNotification
- rename component and introduce QA fixes for button/icon alignment
- update snapshots
