+++ src/components/Toast/index.ts
@@ -1,2 +1,1 @@
 export { Toast as default } from './Toast';
-export { Toast as ToastV2 } from './Toast-v2';
