+++ src/components/ToastNotification/ToastNotification.module.css
@@ -13,7 +13,7 @@
   display: flex;
   gap: 1rem;
   padding: 1rem;
-  align-items: center;
+  align-items: top;
 
   border: 0.125rem solid;
   border-left: 1rem solid;
@@ -37,4 +37,10 @@
 
 .toast__body {
   flex-grow: 2;
+}
+
+.toast__dismiss-button {
+  flex-shrink: 0;
+  margin-top: -0.5rem;
+  margin-right: -0.5rem;
 }
