+++ src/components/ToastNotification/ToastNotification.test.ts
@@ -1,8 +1,8 @@
 import { generateSnapshots } from '@chanzuckerberg/story-utils';
 import type { StoryFile } from '@storybook/testing-react';
 
-import * as stories from './Toast-v2.stories';
+import * as stories from './ToastNotification.stories';
 
-describe('<Toast /> (v2)', () => {
+describe('<ToastNotification /> (v2)', () => {
   generateSnapshots(stories as StoryFile);
 });
