+++ src/components/ToastNotification/index.ts
@@ -1,0 +1,1 @@
+export { ToastNotification as default } from './ToastNotification';
